@@ -1,5 +1,5 @@
 module github.com/woodstock-tokyo/polygon
 
-go 1.17
+go 1.18
 
 require github.com/google/go-querystring v1.1.0
